+++ src/main/java/me/melontini/andromeda/base/util/ToBooleanFunction.java
@@ -1,5 +1,0 @@
-package me.melontini.andromeda.base.util;
-
-public interface ToBooleanFunction<C> {
-    boolean getAsBoolean(C c);
-}