+++ src/main/java/me/melontini/andromeda/modules/blocks/bed/unsafe/Unsafe.java
@@ -6,11 +6,12 @@ import me.melontini.andromeda.base.events.BlockadesEvent;
 import me.melontini.andromeda.base.events.ConfigEvent;
 import me.melontini.andromeda.base.util.Environment;
 import me.melontini.andromeda.base.util.Promise;
-import me.melontini.andromeda.base.util.ToBooleanFunction;
 import me.melontini.andromeda.base.util.annotations.ModuleInfo;
 import me.melontini.andromeda.base.util.config.ConfigDefinition;
 import me.melontini.andromeda.base.util.config.ConfigState;
 import me.melontini.andromeda.modules.blocks.bed.safe.Safe;
+
+import java.util.function.Predicate;
 
 @ModuleInfo(name = "bed/unsafe", category = "blocks", environment = Environment.SERVER)
 public final class Unsafe extends Module {
@@ -19,10 +20,10 @@ public final class Unsafe extends Module {
 
     Unsafe() {
         this.defineConfig(ConfigState.GAME, CONFIG);
-        ToBooleanFunction<ModuleManager> supplier = (manager) -> manager.getDiscovered(Safe.class).map(Promise::get).filter(safe -> manager.getConfig(safe).enabled()).isPresent();
+        Predicate<ModuleManager> supplier = (manager) -> manager.getDiscovered(Safe.class).map(Promise::get).filter(safe -> manager.getConfig(safe).enabled()).isPresent();
 
         ConfigEvent.bootstrap(this).listen((moduleManager, config) -> {
-            if (supplier.getAsBoolean(moduleManager)) {
+            if (supplier.test(moduleManager)) {
                 config.enabled = false;
             }
         });