+++ src/main/java/me/melontini/andromeda/modules/gui/name_tooltips/NameTooltips.java
@@ -5,8 +5,9 @@ import me.melontini.andromeda.base.ModuleManager;
 import me.melontini.andromeda.base.events.BlockadesEvent;
 import me.melontini.andromeda.base.events.ConfigEvent;
 import me.melontini.andromeda.base.util.Environment;
-import me.melontini.andromeda.base.util.ToBooleanFunction;
 import me.melontini.andromeda.base.util.annotations.ModuleInfo;
+
+import java.util.function.Predicate;
 
 import static me.melontini.andromeda.base.Bootstrap.testModVersion;
 
@@ -14,10 +15,10 @@ import static me.melontini.andromeda.base.Bootstrap.testModVersion;
 public final class NameTooltips extends Module {
 
     NameTooltips() {
-        ToBooleanFunction<ModuleManager> iceberg = (manager) -> testModVersion(this, "minecraft", ">=1.20") && testModVersion(this, "iceberg", "<1.1.13");
+        Predicate<ModuleManager> iceberg = (manager) -> testModVersion(this, "minecraft", ">=1.20") && testModVersion(this, "iceberg", "<1.1.13");
 
         ConfigEvent.bootstrap(this).listen((moduleManager, config) -> {
-            if (iceberg.getAsBoolean(moduleManager)) config.enabled = false;
+            if (iceberg.test(moduleManager)) config.enabled = false;
         });
         BlockadesEvent.BUS.listen((manager, blockade) -> {
             blockade.explain(this, "enabled", iceberg, blockade.andromeda("iceberg"));