+++ src/main/java/me/melontini/andromeda/util/CrashHandler.java
@@ -2,13 +2,17 @@ package me.melontini.andromeda.util;
 
 import com.google.common.collect.Sets;
 import com.google.gson.JsonArray;
+import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
+import com.google.gson.JsonPrimitive;
 import lombok.CustomLog;
 import lombok.experimental.UtilityClass;
 import me.melontini.andromeda.base.AndromedaConfig;
+import me.melontini.andromeda.base.Bootstrap;
 import me.melontini.andromeda.util.exceptions.AndromedaException;
 import me.melontini.dark_matter.api.base.util.Context;
 import me.melontini.dark_matter.api.crash_handler.Crashlytics;
+import me.melontini.dark_matter.api.crash_handler.Props;
 import me.melontini.dark_matter.api.crash_handler.uploading.Mixpanel;
 import me.melontini.dark_matter.api.crash_handler.uploading.Uploader;
 import net.fabricmc.loader.api.FabricLoader;
@@ -21,6 +25,7 @@ import java.io.StringWriter;
 import java.nio.charset.StandardCharsets;
 import java.util.*;
 import java.util.concurrent.CompletableFuture;
+import java.util.function.Supplier;
 
 @CustomLog @UtilityClass
 public class CrashHandler {
@@ -29,6 +34,15 @@ public class CrashHandler {
 
     private static final Mixpanel MIXPANEL = Mixpanel.get(new String(Base64.getDecoder().decode("NGQ3YWVhZGRjN2M5M2JkNzhiODRmNDViZWI3Y2NlOTE="), StandardCharsets.UTF_8), true);
     private static final Set<String> IMPORTANT_MODS = Sets.newHashSet("andromeda", "minecraft", "fabric-api", "fabricloader", "connectormod", "forge");
+
+    private static final Map<String, Supplier<JsonElement>> DEFAULT_KEYS = Map.of(
+            "bootstrap_status", () -> new JsonPrimitive(Bootstrap.Status.get().name()),
+            "platform", () -> new JsonPrimitive(CommonValues.platform().toString()),
+            "environment", () -> new JsonPrimitive(Props.ENVIRONMENT.get()),
+            "$os", () -> new JsonPrimitive(Props.OS.get()),
+            "java_version", () -> new JsonPrimitive(Props.JAVA_VERSION.get()),
+            "java_vendor", () -> new JsonPrimitive(Props.JAVA_VENDOR.get())
+    );
 
     private static Flag shouldReportRecursive(Throwable cause, Flag flag) {
         if (cause instanceof AndromedaException e) flag.mark(e.shouldReport());
@@ -74,9 +88,10 @@ public class CrashHandler {
 
     public static void handleCrash(Throwable cause, Context context) {
         if (Debug.Keys.DISABLE_NETWORK_FEATURES.isPresent()) return;
+        var loader = FabricLoader.getInstance();
 
         if (!Debug.Keys.FORCE_CRASH_REPORT_UPLOAD.isPresent()) {
-            if (FabricLoader.getInstance().isDevelopmentEnvironment() || !AndromedaConfig.get().sendCrashReports)
+            if (loader.isDevelopmentEnvironment() || !AndromedaConfig.get().sendCrashReports)
                 return;
         }
 
@@ -89,16 +104,14 @@ public class CrashHandler {
         JsonObject object = new JsonObject();
         //fill trace.
         JsonArray stackTrace = new JsonArray();
-        for (String string : getCauseAsString(cause).lines().flatMap(s -> StringUtil.wrapLines(s, 190).lines()).toList())
-            stackTrace.add(string);
+        getCauseAsString(cause).lines().flatMap(s -> StringUtil.wrapLines(s, 190).lines()).toList().forEach(stackTrace::add);
         object.add("stackTrace", stackTrace);
 
-        object.add("statuses", traverse(cause).orElseGet(AndromedaException::defaultStatuses));
+        DEFAULT_KEYS.forEach((s, supplier) -> object.add(s, supplier.get()));
+        traverse(cause).ifPresent(o -> object.add("statuses", o));
 
-        JsonArray mods = new JsonArray();
-        for (String importantMod : IMPORTANT_MODS) {
-            FabricLoader.getInstance().getModContainer(importantMod).ifPresent(mod -> mods.add(importantMod + " (" + mod.getMetadata().getVersion().getFriendlyString() + ")"));
-        }
+        JsonObject mods = new JsonObject();
+        IMPORTANT_MODS.forEach(importantMod -> loader.getModContainer(importantMod).ifPresent(mod -> mods.addProperty(importantMod,  mod.getMetadata().getVersion().getFriendlyString())));
         object.add("mods", mods);
 
         if (context.get(SKIP_SERVICE).orElse(false)) {
@@ -109,6 +122,12 @@ public class CrashHandler {
                 return null;
             });
         }
+    }
+
+    public static JsonObject dumpState() {
+        JsonObject object = new JsonObject();
+        DEFAULT_KEYS.forEach((s, supplier) -> object.add(s, supplier.get()));
+        return object;
     }
 
     private static void upload(JsonObject object) {